Changed some data types and variable names

+++ data-service/src/main/java/com/axteam/dataservice/dao/DataDao.java
@@ -6,6 +6,6 @@ import java.time.OffsetDateTime;
 import java.util.List;
 
 public interface DataDao {
-	void save(short number, OffsetDateTime time);
+	void save(Integer number, OffsetDateTime time);
 	List<DataRecord> getDataRecord(String startDate, String endDate);
 }
