+++ data-service/src/main/java/com/axteam/dataservice/models/DataRecord.java
@@ -23,12 +23,12 @@ public class DataRecord {
 	private Long id;
 
 	@Column(name = "value")
-	private Short value;
+	private Integer value;
 
 	@Column(name = "time")
 	private OffsetDateTime time;
 
-	public DataRecord(Short value, OffsetDateTime time) {
+	public DataRecord(Integer value, OffsetDateTime time) {
 		this.value = value;
 		this.time = time;
 	}
