+++ data-service/src/main/java/com/axteam/dataservice/services/DataService.java
@@ -5,7 +5,7 @@ import com.axteam.dataservice.models.DataRecord;
 import java.util.List;
 
 public interface DataService {
-	void saveDataRecord(short number);
-	void checkIncident(short number);
+	void saveDataRecord(Integer number);
+	void checkIncident(Integer number);
 	List<DataRecord> getDataRecord(String startDate, String endDate);
 }
